fixes: adding time to logger and fixing the lock logic
* Environment like nix-shells creates a temp_dir for each shell, the old
logic fail to stop the program for running more than one time in these
cases. We are instead placing the lock file at the same place where we
put the database and the log, if it's not possible to put the file there
then the program won't run anyway.
* Logger no more create a new file everytime it starts and now old log
persists and new logs are appended to file with time :D