fix: *probably* preventing a race condition and fixing logic to check idle state
* We were not covering cases where events received by our program from
the wayland compositor were out of order, which would case a race
condition and stop the timer for the new active window.
* Idle check also had a flaw, if user became idle and then came back to
his pc and continued in the same window, the timer would not continue
until user changed to another window.
* We also have a enum for trackingstate now, which helped makes the
logic of recording the active window more clear and easier to read